fix: remediate all 13 findings from senior-SWE code review
Backend (Java):
- Scheduler starvation (#5): add spring.task.scheduling.pool.size and move the
  blocking Alpaca reconciler onto its own executor so a 15s broker call can no
  longer stall strategy/market-making ticks.
- DB write under matching lock (#6): collect fills under synchronized(book), run
  the @Transactional persistence AFTER releasing the lock.
- Money precision (#7): TaxEngine reported dollars -> BigDecimal (exact tax
  figures); RiskEngine/BacktestService keep double (statistical analytics) with a
  documented precision boundary + rounding at the edge.
- Non-atomic LiveOrderBook.resetTo (#8): swap both sides via a single volatile
  holder so readers never see an empty/crossed book.
- Benchmark (#1): the review's "no benchmark" was a false positive
  (MatchingBenchmark.java exists); re-measured ~3.3M ord/s, corrected the noisy
  p99.9 tail in the README and dated it to the test machine.
- Feed parsers (#9): make Coinbase/Alpaca message handling testable + 7 unit
  tests (snapshot/update/trade/auth).
- Live aggregate risk (#10): PreTradeRiskGuard caps book gross notional on the
  live order path (oms.risk.max-gross-notional), not just the backtester.

Research (Python):
- Significance (#3): add sharpe_tstat/sharpe_ci (Lo 2002 SE); run_crosssec/
  run_portfolio now print t-stat + 95% CI. Honest headline: NO signal clears
  |t|>2 (momentum +0.41 t=0.76; only significant result is reversal, a loser).
- Leakage (#12): statarb -> walk-forward OOS (in-sample Sharpe 3.56 collapses to
  -0.62); vol_target made causal (trailing, shifted).
- Minor (#13): relabel _shrink_cov honestly (fixed-lambda, not Ledoit-Wolf);
  crosssec counts active days only (momentum recomputes +0.36 -> +0.41).

Docs & config:
- Reconcile MARKET-REALISM queue-ahead prose to the code (sizeAt, not
  sizeAtOrBetter) (#4); reframe Phase 5/6 with significance; fix stat-arb/test
  counts in README + research/README.
- Pin Python deps (#11); document key rotation in alpaca example (#13).
- "institutional-grade" (#2): not present in the repo; the one
  "institutional capacity question" line is accurate, left as-is.

Verified: backend 105/105 (integration vs real Postgres), research 38/38.

+++ backend/src/main/java/com/bonddesk/oms/tax/dto/TaxDtos.java
@@ -1,5 +1,6 @@
 package com.bonddesk.oms.tax.dto;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -29,34 +30,44 @@ public final class TaxDtos {
     ) {
     }
 
-    /** One realized disposal: which lot, gain/loss, holding period, and any wash-sale disallowance. */
+    /**
+     * One realized disposal: which lot, gain/loss, holding period, and any wash-sale disallowance.
+     * Reported dollar figures ({@code proceeds}, {@code costBasis}, {@code gain}, {@code washDisallowed})
+     * are {@link BigDecimal} at scale 2 so they are exact to the cent; {@code quantity} stays a
+     * double as it is a share/coin count, not money.
+     */
     public record Disposition(
             Instant acquired,
             Instant sold,
             double quantity,
-            double proceeds,
-            double costBasis,
-            double gain,
+            BigDecimal proceeds,
+            BigDecimal costBasis,
+            BigDecimal gain,
             long holdingDays,
             boolean longTerm,
-            double washDisallowed
+            BigDecimal washDisallowed
     ) {
     }
 
+    /**
+     * Tax report. All reported dollar figures are {@link BigDecimal} at scale 2 (cent-exact);
+     * {@code effectiveTaxRate} is a ratio and {@code openPosition}/{@code openAvgBasis} are a
+     * quantity and a per-unit basis, so those stay doubles.
+     */
     public record TaxReport(
             String assetClass,
             String regime,
             String lotMethod,
-            double proceeds,
-            double realizedGain,
-            double shortTermGain,
-            double longTermGain,
-            double washSaleDisallowed,
-            double unrealizedMtm,
-            double taxableGain,
-            double taxOwed,
-            double preTaxPnl,
-            double afterTaxPnl,
+            BigDecimal proceeds,
+            BigDecimal realizedGain,
+            BigDecimal shortTermGain,
+            BigDecimal longTermGain,
+            BigDecimal washSaleDisallowed,
+            BigDecimal unrealizedMtm,
+            BigDecimal taxableGain,
+            BigDecimal taxOwed,
+            BigDecimal preTaxPnl,
+            BigDecimal afterTaxPnl,
             double effectiveTaxRate,
             double openPosition,
             double openAvgBasis,
